+++ LunaFixes/Mods/KerbalPlanetaryBaseSystems/PlanetaryBaseSystemsCompat.cs
@@ -17,7 +17,7 @@ internal class PlanetaryBaseSystemsCompat : ModCompat
 
     #region Public Methods
 
-    public override void Patch(ModMessageHandler modMessageHandler)
+    public override void Patch(ModMessageHandler modMessageHandler, ConfigNode node)
     {
         // TODO add Kerbal Planetary Base Systems fixes here
     }