+++ LunaFixes/Mods/Kethane/KethaneCompat.cs
@@ -23,9 +23,9 @@ internal class KethaneCompat : ModCompat
 
     #region Public Methods
 
-    public override void Patch(ModMessageHandler modMessageHandler)
+    public override void Patch(ModMessageHandler modMessageHandler, ConfigNode node)
     {
-            var legacyResourceGenerator = AccessTools.TypeByName("Kethane.Generators.LegacyResourceGenerator");
+        var legacyResourceGenerator = AccessTools.TypeByName("Kethane.Generators.LegacyResourceGenerator");
 
         LunaFixes.HarmonyInstance.Patch(AccessTools.Method(legacyResourceGenerator, "Load", [
             typeof(CelestialBody), typeof(ConfigNode)