+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file
 ## Unreleased
 
 - add sandcastle compat
+- add SCANsat compat
 
 
 ## 0.0.3 - 2025-07-04